requirements.txt
pygame>=2.5.0
anthropic>=0.74.0
python-dotenv>=1.0.0
